requirements.txt
opencv-python
numpy
Pillow
customtkinter
basicsr
facexlib
gfpgan
realesrgan
requests
tqdm
pyyaml
scipy